Enable arm64 build

+++ shell/platform/tizen/channels/text_input_channel.cc
@@ -277,8 +277,10 @@ TextInputChannel::TextInputChannel(flutter::BinaryMessenger* messenger,
     imf_context_ = ecore_imf_context_add(GetImfMethod());
   }
   if (imf_context_) {
+    // TODO(swift-kim): Incorrect cast from int to void*.
     ecore_imf_context_client_window_set(
-        imf_context_, (void*)engine_->tizen_renderer->GetEcoreWindowId());
+        imf_context_,
+        (void*)(size_t)(engine_->tizen_renderer->GetEcoreWindowId()));
     RegisterIMFCallback();
   } else {
     FT_LOGE("Failed to create imfContext");
